% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivebwdd.R
\name{naivebwdd}
\alias{naivebwdd}
\title{Uses 5-fold cross validation in order to nonparametrically fit linear covariates to directional responses.
hs is a vector of potential bandwidths, x is the linear covariate, y is the directional response, and delta is the points that the model will be fit over.}
\usage{
naivebwdd(hs, x, y, delta)
}
\description{
Uses 5-fold cross validation in order to nonparametrically fit linear covariates to directional responses.
hs is a vector of potential bandwidths, x is the linear covariate, y is the directional response, and delta is the points that the model will be fit over.
}
