% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivedd.R
\name{naivedd}
\alias{naivedd}
\title{Fit linear covariates to directional responses nonparametrically.
h is a chosen bandwidth, x is a linear covariate, y is a directional response, and delta is a vector of points to fit the model over.}
\usage{
naivedd(h, x, y, delta)
}
\arguments{
\item{h}{bandwidth}

\item{x}{covariated}

\item{y}{responses}

\item{delta}{location of regression fit}
}
\description{
Fit linear covariates to directional responses nonparametrically.
h is a chosen bandwidth, x is a linear covariate, y is a directional response, and delta is a vector of points to fit the model over.
}
