% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivebwdd.R
\name{naivebwdd}
\alias{naivebwdd}
\title{Uses SIMEX bandwidth selection in order to fit linear covariates with measurement error to directional responses.
Inputs follow the same as npregdd, except hs is a vector of potential bandwidths.}
\usage{
naivebwdd(hs, x, y, delta)
}
\arguments{
\item{hs}{list of potential bandwidths}

\item{x}{covariates}

\item{y}{responses}

\item{delta}{values for the regression model to be fitted at}
}
\description{
Uses SIMEX bandwidth selection in order to fit linear covariates with measurement error to directional responses.
Inputs follow the same as npregdd, except hs is a vector of potential bandwidths.
}
