% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npddbw.R
\name{npddbw}
\alias{npddbw}
\title{Uses SIMEX bandwidth selection in order to fit linear covariates with measurement error to directional responses.
Inputs follow the same as npregdd, except hs is a vector of potential bandwidths.}
\usage{
npddbw(x, y, hs, method, delta, error, sigmau, rep)
}
\description{
Uses SIMEX bandwidth selection in order to fit linear covariates with measurement error to directional responses.
Inputs follow the same as npregdd, except hs is a vector of potential bandwidths.
}
