% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npregdd.R
\name{npregdd}
\alias{npregdd}
\title{Fits linear covariates with measurement error to directional responses, has three different methods.
h is bandwidth, x is linear covariates, y is directional responses, delta is a vector to fit the model over, sigmau is the standard deviation of the measurement error.
This function has three potential methods, "comp" for complex error correction, "onestep" for onestep error correction, and "deconv" for error correction using deconvoluting kernels.
rep is the number of random normal imaginary errors to generate per replication for complex error correction, increasing this will smooth the fit, but will increase computation time.}
\usage{
npregdd(h, method, x, y, delta, error, sigmau, rep)
}
\arguments{
\item{h}{bandwidth}

\item{method}{"comp" for complex errors "onestep" for one step estimator or "deconv" for deconvoluting estiamtor}

\item{x}{covariates}

\item{y}{responses}

\item{delta}{values for the regression model to be fitted at}

\item{error}{error distribution, "normal" or "laplace"}

\item{sigmau}{standard deviation of the measurement error}

\item{rep}{number of repetitions for monte carlo convergence of complex method}
}
\description{
Fits linear covariates with measurement error to directional responses, has three different methods.
h is bandwidth, x is linear covariates, y is directional responses, delta is a vector to fit the model over, sigmau is the standard deviation of the measurement error.
This function has three potential methods, "comp" for complex error correction, "onestep" for onestep error correction, and "deconv" for error correction using deconvoluting kernels.
rep is the number of random normal imaginary errors to generate per replication for complex error correction, increasing this will smooth the fit, but will increase computation time.
}
